% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_table_to_fasta.R
\name{asv_table_to_fasta}
\alias{asv_table_to_fasta}
\title{asv to fasta}
\usage{
asv_table_to_fasta(asv_table, folder_path = ".", file_name = NULL)
}
\arguments{
\item{asv_table}{An asv table with a column called ASV.}

\item{folder_path}{The path to the folder to store the results}

\item{file_name}{Name of the file, without file extension. Default to \code{asv}.}
}
\description{
This function trasform an ASV table in a fasta file.
}
