% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boldigger.R
\name{boldigger}
\alias{boldigger}
\title{boldigger}
\usage{
boldigger(
  asv_folder = NULL,
  project_path = NULL,
  database_nr = 1,
  operating_mode = 1,
  thresholds = NULL
)
}
\arguments{
\item{asv_folder}{Path to the folder with ASV results.}

\item{project_path}{The path to the project folder.}

\item{database_nr}{Is a number between 1 and 7 corresponding to the seven databases BOLD v5 currently offers.}

\item{operating_mode}{Is a number between 1 and the corresponding to the 3 operating modes BOLD v5 currently offers.}

\item{thresholds}{Threshold for taxonomic identification (defaul Species: 97, Genus: 95, Family: 90, Order: 85)}
}
\description{
This is a wrapper of the python program boldigger3.
}
