% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_asv_table.R
\name{build_asv_table}
\alias{build_asv_table}
\title{ASV table}
\usage{
build_asv_table(project_path = NULL)
}
\arguments{
\item{project_path}{The path to the project folder.}
}
\description{
This function creates an ASV table.
}
