% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_taxonomy.R
\name{build_taxonomy}
\alias{build_taxonomy}
\title{Build BOLD taxonomy}
\usage{
build_taxonomy(folder_path = NULL, asv_table = NULL, query_size = 500)
}
\arguments{
\item{folder_path}{Path to the folder where results are stored.}

\item{asv_table}{An ASV table containing the taxonomic information}

\item{query_size}{Maximum query size to BOLD.}
}
\description{
This function build the taxonomy of species in an ASV table using BOLD.
}
