% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_primers.R
\name{count_primers}
\alias{count_primers}
\title{count primer occurrence}
\usage{
count_primers(project_path = NULL, orientation = "r1")
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{orientation}{The orientation. Default to r1.}
}
\description{
This function prints the number of occurrence in fastq files to the viewer pane.
}
\details{
The information about the number of occurrence is stored in the log_files folder.
}
