% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{count fastq reads}
\usage{
count_reads(project_path = NULL)
}
\arguments{
\item{project_path}{Path to the project folder.}
}
\description{
This function open the log file about read counts stored in the log_files folder.
}
