% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create a project}
\usage{
create_project(project_path)
}
\arguments{
\item{project_path}{The path to the project folder. The folder must already exists.}
}
\description{
This function creates a project to run the metabarcoding pipeline.
}
\details{
The function `create_project()` creates 12 folders in the project folder.
A txt file storing the project path is also created in the log_files folder.
The path should be absolute rather than relative.
}
\examples{
\dontrun{
my_folder_path <- "D:/my_project"
create_project(my_folder_path)
}
}
