% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress_gz_files.R
\name{decompress_gz_file}
\alias{decompress_gz_file}
\title{Decompress gz file}
\usage{
decompress_gz_file(project_path = NULL)
}
\arguments{
\item{project_path}{Path to the project folder.}
}
\description{
Decompress gz files into fastq.
}
