% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Demultiplex multiplexed fastq}
\usage{
demultiplex(
  project_path = NULL,
  r1 = NULL,
  r2 = NULL,
  combinations = NULL,
  indices = NULL,
  folder = NULL
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{r1}{Path to the multiplexed r1 file.}

\item{r2}{Path to the multiplexed r2 file.}

\item{Indices}{used.}

\item{Output}{folder.}
}
\description{
This function demultiplex multiplexed data-
}
