% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{Denoise}
\usage{
denoise(
  project_path = NULL,
  vsearch_path = NULL,
  size_in = TRUE,
  size_out = TRUE,
  minsize = 8,
  unoise_alpha = 2,
  fasta_width = 0,
  vsearch_arguments = NULL,
  n_reads = 0
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{size_in}{From \code{vsearch}: "In de novo mode, abundance annotations present in sequence headers are taken into account by default".}

\item{size_out}{From \code{vsearch}: "...add abundance annotations to fasta headers".}

\item{minsize}{From \code{vsearch}: "Specify the minimum abundance of sequences for denoising using --cluster_unoise. The default is 8".}

\item{unoise_alpha}{From \code{vsearch}: "Specify the alpha parameter to the --cluster_unoise command. The default i 2".}

\item{fasta_width}{From \code{vsearch}: "Fasta files produced by vsearch are wrapped (sequences are written on lines of integer nucleotides, 80 by default). Set that value to zero to eliminate the wrapping".}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.
See \link{names_checker} for further details.}
}
\description{
This function performs denoising.
}
