% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dereplicate.R
\name{dereplicate}
\alias{dereplicate}
\title{Dereplicate FASTQ}
\usage{
dereplicate(
  project_path = NULL,
  vsearch_path = NULL,
  size_out = TRUE,
  size = 1,
  fastq_width = 0,
  relabel = NULL,
  vsearch_arguments = NULL,
  n_reads = 0
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{size_out}{From \code{vsearch}: "...add abundance annotations to fasta headers".}

\item{size}{From \code{vsearch}: "Discard sequences with a post-dereplication abundance value smaller than integer".}

\item{fastq_width}{From \code{vsearch}: "Fasta files produced by vsearch are wrapped (sequences are written on lines of integer nucleotides, 80 by default). Set that value to zero to eliminate the wrapping".}

\item{relabel}{From \code{vsearch}: "Relabel sequences using the prefix string and a ticker (1, 2, 3, etc.) to construct the new headers. Use --sizeout to conserve the abundance annotations".}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.
See \link{names_checker} for further details.}
}
\description{
This function dereplicates FASTQ files to FASTA files.
}
