% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_starting_checks.R
\name{display_starting_checks}
\alias{display_starting_checks}
\title{Display starting checks}
\usage{
display_starting_checks(project_path = NULL)
}
\arguments{
\item{project_path}{Path to the project folder.}
}
\description{
This function display starting checks in the viewer pane.
}
