% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena_request.R
\name{ena_request}
\alias{ena_request}
\title{FASTQ from ENA}
\usage{
ena_request(x = NULL, project_path = NULL)
}
\arguments{
\item{x}{data.frame, basically the tsv downloaded from the ENA website.}

\item{project_path}{The path to the project folder.}
}
\description{
This function download
}
