% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{fastqc}
\usage{
fastqc(project_path = NULL, fastqc_dir = NULL, pattern = "fastq")
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{fastqc_dir}{Path to fastqc directory.}

\item{pattern}{File extension. Default to \code{fastq}.}
}
\value{
The function \code{fastqc} returns quality control files generated by the
program fastqc in the `reports` folder.
}
\description{
This function is a wrapper of the software FASTQC.
}
\details{
The function \code{fastqc} uses the program fastqc to do some quality control checks.
The fastqc program is available at \url{https://www.bioinformatics.babraham.ac.uk/projects/fastqc/}.
This function requires Java and FastQC to be installed and accessible.
Java must be available on the system PATH.
}
