% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_rename.R
\name{file_rename}
\alias{file_rename}
\title{rename fastq files}
\usage{
file_rename(
  project_path = NULL,
  metadata = NULL,
  file_folder = NULL,
  r1_original = NULL,
  r2_original = NULL,
  r1_new = "_r1.fastq",
  r2_new = "_r2.fastq"
)
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{metadata}{A \code{data.frame} or \code{tibble} to change fastq names. The column
of the old names must be named \code{fastq}, the column for the new names \code{fastq_new}.}

\item{file_folder}{The folder containing r1 and r2 files.}

\item{r1_original}{The pattern to replace for r1.}

\item{r2_original}{The pattern to replace for r2.}

\item{r1_new}{The new pattern for r1.}

\item{r2_new}{The new pattern for r2.}
}
\description{
This function rename files based on specific patterns. It is intended to rename
raw fastq files to fit with the pipeline.
}
\details{
The pipeline needs input fastqs end with the pattern r1.fastq. The function
`create_project()`try to do this for input raw files, that must be fastq.
}
