% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_quality.R
\name{filter_by_quality}
\alias{filter_by_quality}
\title{Filter by quality}
\usage{
filter_by_quality(
  project_path = NULL,
  vsearch_path = NULL,
  output_path = NULL,
  max_ee = 1,
  qmax = 41,
  vsearch_arguments = NULL,
  n_reads = 0,
  fasta_out = TRUE
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{output_path}{The path were saving the results.}

\item{max_ee}{From \code{vsearch}: "...discard sequences with more than the specified number of expected errors"}

\item{qmax}{From \code{vsearch}: "...the maximum quality score accepted when reading FASTQ files. Default to 41."}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.
See \link{names_checker} for further details.}

\item{fasta_out}{If \code{TRUE} return a fasta otherwise a fastq.}
}
\description{
This function filter by quality using the expected error (ee) approach.
}
