% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tag_combinations.R
\name{find_tag_combinations}
\alias{find_tag_combinations}
\title{Find tag combinations in multiplexed fastq}
\usage{
find_tag_combinations(r1, r2, primer_fwd, primer_rvr, project_path)
}
\arguments{
\item{r1}{Path to the multiplexed r1 file.}

\item{r2}{Path to the multiplexed r2 file.}

\item{primer_fwd}{Forward primers.}

\item{primer_rvr}{Reverse primers.}

\item{project_path}{The path to the project folder.}
}
\description{
This function tries to infer tags from the multiplexed fastq.
}
\details{
This function starts by filtering the multiplexed fastq by an expected
error of 3 and a sequence length of 300. The cut the first 6 bases and
count the frequency of the strings in the overall files. A summary file
and the inferred indices are written to disk.
}
