% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make a fastq table}
\usage{
make_table(x)
}
\arguments{
\item{x}{A table downloaded from ENA.}
}
\description{
This function make a sample table from e ENA table.
}
