% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replicates.R
\name{merge_replicates}
\alias{merge_replicates}
\title{Merge sample replicates}
\usage{
merge_replicates(
  project_path = NULL,
  asv_table = "translated",
  pattern = "_replicate",
  percentage_to_remove = 0
)
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{asv_table}{Can be \code{chimera_removed} for the asv_table generated after
chimera removal or \code{chimera} for the asv_table generated after protein translation.}

\item{pattern}{The pattern to remove. See details.}

\item{percentage_to_remove}{Remove ASV with a relative abundance lower than this threshold.}
}
\description{
This function merge multiple replicates of the same sample.
}
\details{
Replicates of the same sample are usually stored with the name of the sample and
an indication of the replicate (e.g. replicate1, replicate2). Please provide a pattern
without the replicate number (e.g. \code{pattern = "replicate"}).
}
