% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc}
\alias{multiqc}
\title{Run multiqc}
\usage{
multiqc(project_path = NULL)
}
\arguments{
\item{project_path}{Path to the project folder.}
}
\value{
The function `multiqc`return an html file that is open in the default browser
}
\description{
This function run the python progam multiqc.
}
\details{
The function `multiqc` run the python program multiqc. Python and multiqc must
be already installed. See \url{https://multiqc.info/} for more details.
}
