% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_checker.R
\name{names_checker}
\alias{names_checker}
\title{Check fastq names from r1 and r2}
\usage{
names_checker(
  project_path = NULL,
  file_folder = NULL,
  r1_pattern = "_r1.fastq",
  r2_pattern = "_r2.fastq",
  return_fastq_names = FALSE,
  n_reads = 0
)
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{file_folder}{The folder containing r1 and r2 files.}

\item{r1_pattern}{Pattern to remove from r1 to make the comparison with r2.
Default to `_r1.fastq.`}

\item{r2_pattern}{Pattern to remove from r2 to make the comparison with r1.
Default to `_r2.fastq.`}

\item{return_fastq_names}{if \code{TRUE}, returns a list of fastq files present
in both r1 and r2.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.}
}
\value{
The function `r1_r2_checker` returns TRUE if all the r1 files are matched by
files in r2, otherwise FALSE. If \code{return_fastq_names} is set to \code{TRUE},
the function returns a list of r1 and r2 fastq.
}
\description{
This function checks if r1 and r2 files match.
}
\details{
The function `r1_r2_checker` remove patterns from r1 and r2 to check if the
file names from r1 and r2 are the same.
}
