% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_multiqc.R
\name{open_multiqc}
\alias{open_multiqc}
\title{open multiqc report}
\usage{
open_multiqc(project_path = NULL)
}
\arguments{
\item{project_path}{Path to the project folder.}
}
\description{
This function open the report generated by multiqc
}
