% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_clustering.R
\name{otu_clustering}
\alias{otu_clustering}
\title{OTU clustering with vsearch}
\usage{
otu_clustering(
  asv_folder = "13_taxonomic_assignment",
  project_path = NULL,
  vsearch_path = NULL,
  id = 0.97,
  vsearch_arguments = NULL
)
}
\arguments{
\item{asv_folder}{Path to the folder with ASV results.}

\item{project_path}{Path to the project folder.}

\item{vsearch_path}{Path to vsearch.}

\item{id}{From \code{vsearch}: "The pairwise identity is defined as the number of (matching columns) / (alignment length - terminal gaps)".}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}
}
\description{
This function perferms the OTU clustering with \code{vsearch}.
}
