% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_end_merge.R
\name{paired_end_merge}
\alias{paired_end_merge}
\title{Paired end merge}
\usage{
paired_end_merge(
  project_path = NULL,
  vsearch_path = NULL,
  fastq_maxdiffs = 10,
  fastq_maxdiffpct = 100,
  fastq_minovlen = 10,
  mergestagger = TRUE,
  vsearch_arguments = NULL
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{fastq_maxdiffs}{From \code{vsearch}: "... specify the maximum number of non-matching nucleotides allowed in the overlap region. The default value is 10."}

\item{fastq_maxdiffpct}{From \code{vsearch}: "..specify the maximum percentage of non-matching nucleotides allowed in the overlap region. The default value is 100.0 \%."}

\item{fastq_minovlen}{From \code{vsearch}: "...specify the minimum overlap between the merged reads. The default is 10".}

\item{mergestagger}{From \code{vsearch}: "...allow to merge staggered read pairs. Staggered pairs
are pairs where the 3' end of the reverse read has an overhang to the left of the 5' end
 of the forward read."}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}
}
\description{
This function performs paired end merging.
}
