% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_orientation.R
\name{primer_orientation}
\alias{primer_orientation}
\title{Quantify the primer orientation}
\usage{
primer_orientation(
  project_path = NULL,
  primer_fwd = NULL,
  primer_rvr = NULL,
  n = 10000
)
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{primer_fwd}{Forward primer.}

\item{primer_rvr}{Reverse primer.}

\item{n}{The number of reads to subset for calculating primer occurrence.
Default to 1e4.}
}
\description{
This function assess the primer forms. this is useful to take the reverse
of the complement in a later step.
}
\details{
The function \code{primer_orientation} compare the read names of r1 and r2 on a
random subset of the reads. Increase \code{n} if you need more accurate results.
}
