% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimera_removal.R
\name{remove_chimera}
\alias{remove_chimera}
\title{Chimera removal}
\usage{
remove_chimera(
  project_path,
  vsearch_path,
  size_in = TRUE,
  size_out = TRUE,
  fasta_width = 0,
  relabel = NULL,
  vsearch_arguments = NULL,
  n_reads = 0
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{size_in}{From \code{vsearch}: "In de novo mode, abundance annotations present in sequence headers are taken into account by default".}

\item{size_out}{From \code{vsearch}: "...add abundance annotations to fasta headers".}

\item{fasta_width}{From \code{vsearch}: "Fasta files produced by vsearch are wrapped (sequences are written on lines of integer nucleotides, 80 by default). Set that value to zero to eliminate the wrapping".}

\item{relabel}{From \code{vsearch}: "Relabel sequences using the prefix string and a ticker (1, 2, 3, etc.) to construct the new headers. Use --sizeout to conserve the abundance annotations".}

\item{vsearch_arguments}{Further arguments to be passed to \code{vsearch}.}

\item{n_reads}{The number of reads above which a fasta is kept for further processing.}
}
\description{
This function perfroms chimera removal.
}
