% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Fastq reverse complement}
\usage{
reverse_complement(project_path = NULL, primer_fwd = NULL, vsearch_path = NULL)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{primer_fwd}{Forward primer.}

\item{vsearch_path}{The path to \code{vsearch} folder.}
}
\description{
This function compute the reverse complement of a fastq.
}
\details{
The function \code{reverse_complement} replace the original file with the reverse
complement. The result is irreversible.
}
