% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_checks.R
\name{starting_checks}
\alias{starting_checks}
\title{Further quality checks}
\usage{
starting_checks(
  project_path = NULL,
  r1_pattern = "_r1.fastq",
  r2_pattern = "_r2.fastq",
  n = 10000
)
}
\arguments{
\item{project_path}{Path to the project folder.}

\item{r1_pattern}{The pattern of r1. Default to \code{_r1.fastq}.}

\item{r2_pattern}{The pattern of r1. Default to \code{_r2.fastq}.}

\item{n}{The number of reads to compare the read names of r1 and r2.}
}
\description{
This function compare the number of reads and the reads name between
r1 and r2
}
\details{
The function \code{starting checks} compare the read names of r1 and r2 on a
subset of the reads. Increase \code{n} if you need more accurate results.
}
