% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm.R
\name{swarm}
\alias{swarm}
\title{OTU clustering with swarm}
\usage{
swarm(
  asv_table = NULL,
  swarm_path = NULL,
  param_d = 13,
  param_t = 4,
  swarm_arguments = NULL
)
}
\arguments{
\item{asv_table}{ASV table.}

\item{swarm_path}{Path to swarm.}

\item{param_d}{Distance parameter.}

\item{param_t}{Parameter t.}

\item{swarm_arguments}{Further arguments to be passed to \code{swarm}.}
}
\description{
This function perfroms the OTU clustering with \code{swarm}
}
