% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_assignment.R
\name{taxonomic_assignment}
\alias{taxonomic_assignment}
\title{taxonomic assignment}
\usage{
taxonomic_assignment(
  project_path = NULL,
  asv_folder = NULL,
  species_only = FALSE,
  remove_bad_names = TRUE,
  filter_taxlev = NULL,
  filter_taxa = NULL
)
}
\arguments{
\item{project_path}{Path where searching for boldigger results}

\item{asv_folder}{Path to the folder with ASV results.}

\item{species_only}{Keep records with species name only.}

\item{remove_bad_names}{Remove badly formatted species names.}

\item{filter_taxlev}{Select a taxonomic level to subset (e.g. Phylum)}

\item{filter_taxa}{Select one taxon or more taxa to filter at the selected taxonomic level.}
}
\description{
taxonomic assignment
}
\details{
This function build a table using boldigger dta
}
