% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_sequences.R
\name{translate_sequences}
\alias{translate_sequences}
\title{Translation filter}
\usage{
translate_sequences(
  project_path = NULL,
  m = NULL,
  M = NULL,
  genetic_code = NULL,
  check_reverse = FALSE
)
}
\arguments{
\item{project_path}{The path to the project folder. The folder must already exists.}

\item{m}{Minimum marker length}

\item{M}{Maximum marker length}

\item{genetic_code}{Genetic code for sequences translation.}

\item{check_reverse}{Also check reverse-complement reading frames? Default to \code{FALSE}}
}
\description{
This function select sequences that do not have a stop codon.
}
