% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_fastq.R
\name{trim_fastq_length}
\alias{trim_fastq_length}
\title{Trim by length}
\usage{
trim_fastq_length(
  project_path = NULL,
  vsearch_path = NULL,
  usearch_path = NULL,
  n_reads = 0,
  trim_r1 = NULL,
  trim_r2 = NULL,
  left_r1 = 0,
  right_r1 = 0,
  left_r2 = 0,
  right_r2 = 0,
  min_length_r1 = NULL,
  min_length_r2 = NULL,
  vsearch_arguments_r1 = NULL,
  vsearch_arguments_r2 = NULL
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{vsearch_path}{The path to \code{vsearch} folder.}

\item{usearch_path}{The path to \code{usearch} folder.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.}

\item{trim_r1}{Trim length of r1.}

\item{trim_r2}{Trim length of r2.}

\item{left_r1}{Number of bases to remove on the left r1.}

\item{right_r1}{Number of bases to remove on the right r1.}

\item{left_r2}{Number of bases to remove on the left r2.}

\item{right_r2}{Number of bases to remove on the right r2.}

\item{min_length_r1}{Minum length of the reads to keep for r1.}

\item{min_length_r2}{Minum length of the reads to keep for r2.}

\item{vsearch_arguments_r1}{Further arguments to be passed to \code{vsearch} for r1.}

\item{vsearch_arguments_r2}{Further arguments to be passed to \code{vsearch} for r2.}
}
\description{
This function trim the left or the right side of the reads in the raw fastq.
}
\details{
The function \code{trim_fastq_length} relies on \code{vsearch}. It is intended
for trimming raw fastq files. Sometimes the reads from one orientation can be
trimmed more than the other, resulting in a mismatch between reads ID of the
two orientations. When a mismatch is detected, the function \code{trim_fastq_length}
perform a simultaneus trimming between the two orientation. With the simultaneous option
do not allow different trimming between r1 and r2.  This behavior will likely
change in the future.
}
