% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_primers.R
\name{trim_primers}
\alias{trim_primers}
\title{Trim primers}
\usage{
trim_primers(
  project_path = NULL,
  cutadapt_path = NULL,
  primer_fwd = NULL,
  primer_rvr = NULL,
  m = 0,
  n_cores = 1,
  n_times = 1,
  n_errors = 0,
  discard_untrimmed = TRUE,
  cutadapt_arguments = NULL
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{cutadapt_path}{The path to \code{cutadapt} folder.}

\item{primer_fwd}{Forward primer.}

\item{primer_rvr}{Reverse primer.}

\item{m}{From \code{cutadapt}: "... discard processed reads that are shorter than LENGTH".}

\item{n_cores}{The number of cores.}

\item{n_times}{From \code{cutadapt}: "... number of times an adapter is removed on the same sequence".}

\item{n_errors}{The number of allowed errors.}

\item{discard_untrimmed}{From \code{cutadapt}: "... --pair-filter=any with --discard-untrimmed, the pair is discarded if one of the reads does not contain an adapter".}

\item{cutadapt_arguments}{Further arguments to be passed to \code{cutadapt}.}
}
\description{
This function trims primers from merged reads.
}
\details{
See cutadapt website \url{https://cutadapt.readthedocs.io/en/stable/guide.html}.
}
