% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_to_marker_length.R
\name{trim_to_marker_length}
\alias{trim_to_marker_length}
\title{Trim to marker length}
\usage{
trim_to_marker_length(
  project_path = NULL,
  cutadapt_path = NULL,
  m = NULL,
  M = NULL,
  n_cores = 1,
  n_reads = 0
)
}
\arguments{
\item{project_path}{The path to the project folder.}

\item{cutadapt_path}{The path to \code{cutadapt} folder.}

\item{m}{Minimum length.}

\item{M}{Maximum length.}

\item{n_cores}{The number of cores.}

\item{n_reads}{The number of reads above which a fastq is kept for further processing.
See \link{names_checker} for further details.}
}
\description{
Thi function trim reads to marker length.
}
